removed duplicate run_tests.sh from container, split model tests into two files, improved github action naming
