+++ .github/actions/test_action/action.yml
@@ -1,6 +1,6 @@
-name: 'CI_action'
-description: 'Runs unit tests upon push.'
+name: 'unittest docker action'
+description: 'Creates docker cotainer and runs unittests upon push.'
 runs:
   using: 	'docker'
   image: 	'./../../../Dockerfile'
-  entrypoint:	"/home/jovyan/run_tests.sh"
+  entrypoint:	"/home/jovyan/work/run_tests.sh"
