+++ Dockerfile
@@ -24,6 +24,3 @@ ADD . /home/jovyan/work
 WORKDIR /home/jovyan/work
 RUN pip install .
 
-# to run tests
-COPY run_tests.sh /home/jovyan/run_tests.sh
-
