+++ run_test_docker.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-docker run -it --entrypoint '/home/jovyan/run_tests.sh' eu.gcr.io/unit8-product/u8timeseries:latest
+docker run -it --entrypoint '/home/jovyan/work/run_tests.sh' eu.gcr.io/unit8-product/u8timeseries:latest
